requirements.txt
appnope==0.1.0
backcall==0.1.0
cycler==0.10.0
decorator==4.4.2
imagecodecs==2020.5.30
imageio==2.8.0
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
kiwisolver==1.2.0
matplotlib==3.2.1
networkx==2.4
numpy==1.18.5
opencv-python==4.2.0.34
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.4.1
setuptools==45.3.0
six==1.15.0
tifffile==2020.5.30
traitlets==4.3.3
wcwidth==0.1.9
openslide-python==1.1.1

